+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.0] - 05/11/2025
+
+### Added
+ - Added the search for external collabs, allowing to search by name or lastname.
+ - Added filters to obtain external collabs by role or status.
+ - Added the possibility to order external collabs by name ASC or DESC.
+
  ## [0.0.0] - 04/11/2025 
 
 ### Added